Moved mocks to appropiate place. Removed mocks file
