--- core-web/libs/portlets/dot-ema/src/lib/edit-ema-editor/components/edit-ema-palette/shared/edit-ema-palette.mocks.ts
@@ -1,165 +1,0 @@
-import {
-    DotPaletteState,
-    EditEmaPaletteStoreStatus,
-    PALETTE_TYPES
-} from '../store/edit-ema-palette.store';
-
-export const CONTENT_TYPE_MOCK = [
-    {
-        name: 'Test Content Type',
-        variable: 'Test1',
-        icon: 'icon'
-    },
-    {
-        name: 'Test Content Type 2',
-        variable: 'Test2',
-        icon: 'icon'
-    }
-];
-
-export const CONTENTLETS_MOCK = [
-    {
-        hostName: 'demo.dotcms.com',
-        modDate: '2021-04-08 13:53:32.618',
-        imageMetaData: {
-            modDate: 1703020595125,
-            sha256: '01bed04a0807b45245d38188da3bece44e42fcdd0cf8e8bfe0585e8bd7a61913',
-            length: 15613,
-            title: 'box-info-2-270x270.jpg',
-            version: 20220201,
-            isImage: true,
-            fileSize: 15613,
-            name: 'box-info-2-270x270.jpg',
-            width: 270,
-            contentType: 'image/jpeg',
-            height: 270
-        },
-        publishDate: '2021-04-08 13:53:32.681',
-        description:
-            "Snowboarding, once a prime route for teen rebellion, today is definitely mainstream. Those teens — both guys and Shred Bettys, who took up snowboarding in the late '80s and '90s now are riding with their kids.",
-        title: 'Snowboarding',
-        body: "<p>As with skiing, there are different styles of riding. Free-riding is all-mountain snowboarding on the slopes, in the trees, down the steeps and through the moguls. Freestyle is snowboarding in a pipe or park filled with rails, fun boxes and other features.<br /><br />Snowboarding parks are designed for specific skill levels, from beginner parks with tiny rails hugging the ground to terrain parks with roller-coaster rails, fun boxes and tabletops for more experienced snowboarders.<br /><br />Whether you're a first-timer or already comfortable going lip-to-lip in a pipe, there are classes and special clinics for you at our ski and snowboard resorts. Our resorts offer multiday clinics, so if you're headed to ski this winter, consider wrapping your vacation dates around a snowboarding clinic.</p>",
-        baseType: 'CONTENT',
-        inode: 'd77576ce-6e3a-4cf3-b412-8e5209f56cae',
-        archived: false,
-        host: '48190c8c-42c4-46af-8d1a-0cd5db894797',
-        working: true,
-        locked: false,
-        stInode: '778f3246-9b11-4a2a-a101-e7fdf111bdad',
-        contentType: 'Activity',
-        live: true,
-        altTag: 'Snowboarding',
-        owner: 'dotcms.org.1',
-        imageVersion: '/dA/d77576ce-6e3a-4cf3-b412-8e5209f56cae/image/box-info-2-270x270.jpg',
-        identifier: '574f0aec-185a-4160-9c17-6d037b298318',
-        image: '/dA/574f0aec-185a-4160-9c17-6d037b298318/image/box-info-2-270x270.jpg',
-        imageContentAsset: '574f0aec-185a-4160-9c17-6d037b298318/image',
-        urlTitle: 'snowboarding',
-        languageId: 1,
-        URL_MAP_FOR_CONTENT: '/activities/snowboarding',
-        url: '/content.2f6fe5b8-a2cc-4ecb-a868-db632d695fca',
-        tags: 'snowboarding,winterenthusiast',
-        titleImage: 'image',
-        modUserName: 'Admin User',
-        urlMap: '/activities/snowboarding',
-        hasLiveVersion: true,
-        folder: 'SYSTEM_FOLDER',
-        hasTitleImage: true,
-        sortOrder: 0,
-        modUser: 'dotcms.org.1',
-        __icon__: 'contentIcon',
-        contentTypeIcon: 'paragliding',
-        variant: 'DEFAULT'
-    },
-    {
-        hostName: 'demo.dotcms.com',
-        modDate: '2020-09-02 16:42:10.049',
-        imageMetaData: {
-            modDate: 1703020594791,
-            sha256: '7e1cf9d3c8c144f592af72658456031c8283bfe4a5ecce3e188c71aa7b1e590e',
-            length: 37207,
-            title: 'zip-line.jpg',
-            version: 20220201,
-            isImage: true,
-            fileSize: 37207,
-            name: 'zip-line.jpg',
-            width: 270,
-            contentType: 'image/jpeg',
-            height: 270
-        },
-        publishDate: '2020-09-02 16:42:10.101',
-        description:
-            'Ever wondered what it is like to fly through the forest canopy? Zip-lining ais the best way to explore the forest canopy, where thick branches serve as platforms for the adventurous traveler, more than 100 feet above the forest floor.',
-        title: 'Zip-Lining',
-        body: '<p>Ever wondered what it is a monkey finds so fascinating about the forest canopy? Costa Rica is a pioneer in canopy exploration, where thick branches serve as platforms for the adventurous traveler, more than 100 feet above the jungle floor. If you&rsquo;re wondering why you&rsquo;d want to head to the top of a tree just to look around, remember that 90% of Costa Rica animals and 50% of plant species in rainforests live in the upper levels of the trees. When you explore that far off the ground, the view is something you&rsquo;ll never forget! A Costa Rica zip line tour, hanging bridges hike, or aerial tram tour are all fantastic ways to take advantage of Costa Rica&rsquo;s stunning forest canopy views.</p>\n<p>Almost anyone of any age and physical condition can enjoy a Costa Rica zip line adventure as it is not strenuous. Secured into a harness and attached to a sturdy cable, you will have the opportunity to fly through the rainforest canopy and experience a bird&rsquo;s eye view of the lively forest below. A Costa Rica zip line is about a five hour adventure that operates rain or shine all year and is led by bilingual guides. For the non-adrenaline junkie, the aerial tram can take you through the rainforest in comfort and safety. Prefer to linger? Hanging bridges offer panoramic views for acres, and an experienced guide will be happy to point out a variety of birds and animals.</p>',
-        baseType: 'CONTENT',
-        inode: '8df9a375-0386-401c-b5d6-da21c1c5c301',
-        archived: false,
-        host: '48190c8c-42c4-46af-8d1a-0cd5db894797',
-        working: true,
-        locked: false,
-        stInode: '778f3246-9b11-4a2a-a101-e7fdf111bdad',
-        contentType: 'Activity',
-        live: true,
-        altTag: 'Zip-line',
-        owner: 'dotcms.org.1',
-        imageVersion: '/dA/8df9a375-0386-401c-b5d6-da21c1c5c301/image/zip-line.jpg',
-        identifier: '50757fb4-75df-4e2c-8335-35d36bdb944b',
-        image: '/dA/50757fb4-75df-4e2c-8335-35d36bdb944b/image/zip-line.jpg',
-        imageContentAsset: '50757fb4-75df-4e2c-8335-35d36bdb944b/image',
-        urlTitle: 'zip-lining',
-        languageId: 1,
-        URL_MAP_FOR_CONTENT: '/activities/zip-lining',
-        url: '/content.c17f9c4c-ad14-4777-ae61-334ee6c9fcbf',
-        tags: 'ecoenthusiast,zip-lining',
-        titleImage: 'image',
-        modUserName: 'Admin User',
-        urlMap: '/activities/zip-lining',
-        hasLiveVersion: true,
-        folder: 'SYSTEM_FOLDER',
-        hasTitleImage: true,
-        sortOrder: 0,
-        modUser: 'dotcms.org.1',
-        __icon__: 'contentIcon',
-        contentTypeIcon: 'paragliding',
-        variant: 'DEFAULT'
-    }
-];
-
-export const INITIAL_STATE_PALETTE_CONTENTTYPE_MOCK: DotPaletteState = {
-    contentlets: {
-        filter: {
-            query: '',
-            contentTypeVarName: ''
-        },
-        items: [],
-        totalRecords: 0,
-        itemsPerPage: 10
-    },
-    contenttypes: {
-        items: [],
-        filter: ''
-    },
-    status: EditEmaPaletteStoreStatus.LOADED,
-    currentPaletteType: PALETTE_TYPES.CONTENTTYPE,
-    allowedTypes: []
-};
-
-export const INITIAL_STATE_PALETTE_CONTENTLET_MOCK: DotPaletteState = {
-    contentlets: {
-        filter: {
-            query: '',
-            contentTypeVarName: 'Test'
-        },
-        items: CONTENTLETS_MOCK,
-        totalRecords: CONTENTLETS_MOCK.length,
-        itemsPerPage: 10
-    },
-    contenttypes: {
-        items: [],
-        filter: ''
-    },
-    status: EditEmaPaletteStoreStatus.LOADED,
-    currentPaletteType: PALETTE_TYPES.CONTENTLET,
-    allowedTypes: []
-};
